@@ -1,10 +1,10 @@
-﻿using Newtonsoft.Json;
+﻿using System.Text.Json.Serialization;
 
 namespace HenBot;
 
 public record PostObject
 {
-    [JsonProperty("@attributes")] public Attributes Attributes { get; set; }
+    [JsonPropertyName("@attributes")] public Attributes Attributes { get; set; }
 
-    [JsonProperty("post")] public List<Post> Post { get; set; }
+    [JsonPropertyName("post")] public List<Post> Post { get; set; }
 }
