@@ -1,10 +1,10 @@
-﻿using Newtonsoft.Json;
+﻿using System.Text.Json.Serialization;
 
 namespace HenBot;
 
 public record TagObject
 {
-    [JsonProperty("@attributes")] public Attributes Attributes { get; set; }
+    [JsonPropertyName("@attributes")] public Attributes Attributes { get; set; }
 
     public List<Tag> Tag { get; set; }
 }
