@@ -6,7 +6,7 @@ public static class StartHandler
 {
     public static async Task HandleStart(ITelegramBotClient botClient, long chatId, CancellationToken cancellationToken)
     {
-        await botClient.SendTextMessageAsync(
+        await botClient.SendMessage(
             chatId,
             "Write /settings to initial configuration or /getAyaya to get anime pics",
             cancellationToken: cancellationToken);
